fixed review suggestions
