+++ rectools/model_selection/time_split.py
@@ -19,11 +19,11 @@ from datetime import date, datetime
 
 import numpy as np
 import pandas as pd
-from scipy import sparse
 
 from rectools import Columns
 from rectools.dataset import Interactions
-from rectools.utils import isin_2d_int, pairwise
+from rectools.model_selection.utils import get_not_seen_mask
+from rectools.utils import pairwise
 
 DateRange = tp.Sequence[tp.Union[date, datetime]]
 
@@ -186,60 +186,3 @@ class TimeRangeSplitter:
     @staticmethod
     def _get_real_date_range(series_datetime: pd.Series, date_range: DateRange) -> pd.Series:
         return date_range[(date_range >= series_datetime.min()) & (date_range <= series_datetime.max())]
-
-
-def get_not_seen_mask(
-    train_users: np.ndarray,
-    train_items: np.ndarray,
-    test_users: np.ndarray,
-    test_items: np.ndarray,
-) -> np.ndarray:
-    """
-    Return mask for test interactions that is not in train interactions.
-
-    Parameters
-    ----------
-    train_users : np.ndarray
-        Integer array of users in train interactions (it's not a unique users!).
-    train_items : np.ndarray
-        Integer array of items in train interactions. Has same length as `train_users`.
-    test_users : np.ndarray
-        Integer array of users in test interactions (it's not a unique users!).
-    test_items : np.ndarray
-        Integer array of items in test interactions. Has same length as `test_users`.
-
-    Returns
-    -------
-    np.ndarray
-     Boolean mask of same length as `test_users` (`test_items`).
-     ``True`` means interaction not present in train.
-    """
-    if train_users.size != train_items.size:
-        raise ValueError("Lengths of `train_users` and `train_items` must be the same")
-    if test_users.size != test_items.size:
-        raise ValueError("Lengths of `test_users` and `test_items` must be the same")
-
-    if train_users.size == 0:
-        return np.ones(test_users.size, dtype=bool)
-    if test_users.size == 0:
-        return np.array([], dtype=bool)
-
-    n_users = max(train_users.max(), test_users.max()) + 1
-    n_items = max(train_items.max(), test_items.max()) + 1
-
-    cls = sparse.csr_matrix if n_users < n_items else sparse.csc_matrix
-
-    def make_matrix(users: np.ndarray, items: np.ndarray) -> sparse.spmatrix:
-        return cls((np.ones(len(users), dtype=bool), (users, items)), shape=(n_users, n_items))
-
-    train_mat = make_matrix(train_users, train_items)
-    test_mat = make_matrix(test_users, test_items)
-
-    already_seen_coo = test_mat.multiply(train_mat).tocoo(copy=False)
-    del train_mat, test_mat
-    already_seen_arr = np.vstack((already_seen_coo.row, already_seen_coo.col)).T.astype(test_users.dtype)
-    del already_seen_coo
-
-    test_ui = np.vstack((test_users, test_items)).T
-    not_seen_mask = isin_2d_int(test_ui, already_seen_arr, invert=True)
-    return not_seen_mask
